==> NCMF/ncmf_sim_study/1 - syn_data_gen - generate required datasets - try 1 - final.ipynb <==
# real and binary matrix generation + sparsity + noise
# + refactored
# + code to generate the needed datasets

import numpy as np
np.random.seed(0)
import pandas as pd
import os
from datetime import datetime
import time
from sklearn.preprocessing import MaxAbsScaler, StandardScaler, maxabs_scale
from math import pi
from numpy import sin, cos, linspace
import random
import seaborn as sns
import pandas as pd
import matplotlib.pyplot as plt
import pickle as pkl

from ncmf_sim_data_generator import SyntheticDataGeneratorBioinf

import bz2
import _pickle as cPickle

#https://betterprogramming.pub/load-fast-load-big-with-compressed-pickles-5f311584507e

# Pickle a file and then compress it into a file with extension 
def compressed_pickle(fname, data):
    assert fname.endswith("pbz2")
    with bz2.BZ2File(fname, "w") as f: 
        cPickle.dump(data, f)

 # Load any compressed pickle file
def decompress_pickle(file):
    data = bz2.BZ2File(file, "rb")
    data = cPickle.load(data)
    return data

k = 100
G = {"0":["0","1"],"1":["0","2"],"2":["0","3"],"3":["4","1"],"4":["5","2"]}
dict_entity_size = {'0':200, '1':500, '2':300, '3':400, '4':700, '5':600}
dict_mat_dtype = {'0':"real", '1':"real", '2':"binary", '3':"real", '4':"binary"}
#key: matrix index, 
#value: list; where,
#list[0] - row entity index, list[1] - col
#
list_mod_mat_id = ["0","2"] # we modify the sparsity and the noise level only to these matrices
#sparsity_level = 0
# noise_level = 1
# 0 - no modification to the sparsity/noise
# 1 to 3 - increase in the number => increase in noise/sparsity level i.e. number of zeros

#generate datasets by varying sparsity_level and noise_level

dict_name_dataset = {}

#dataset 1
#dt1 - aug multiview dataset with matrices of multiple data types

sparsity_level = 0
noise_level = 0
dname = "dt1"

data_genrator = SyntheticDataGeneratorBioinf(k, G, dict_entity_size, dict_mat_dtype,\
                                                list_mod_mat_id, sparsity_level, noise_level)
data_dict = data_genrator.get_data_dict()

dict_name_dataset[dname] = data_dict

#

#dataset 2
#ds1, ds2, ds3 - aug multiview dataset with matrices of multiple data types + varying sparsity
#sparsity(ds1) < sparsity(ds2) , sparsity(ds3)
sparsity_level_list = [1,2,3]
for sparsity_level in sparsity_level_list:
    noise_level = 0
    dname = "ds"+str(sparsity_level)
    #
    print("###")
    print("dataset: ",dname)
    print("###")
    #
    data_genrator = SyntheticDataGeneratorBioinf(k, G, dict_entity_size, dict_mat_dtype,\
                                                list_mod_mat_id, sparsity_level, noise_level)
    data_dict = data_genrator.get_data_dict()
    #
    dict_name_dataset[dname] = data_dict

#dataset 3
#dn1, dn2, dn3 - aug multiview dataset with matrices of multiple data types + varying noise levels
#noise(dn1) < noise(dn2) , noise(dn3)
noise_level_list = [1,2,3]
for noise_level in noise_level_list:
    sparsity_level = 0
    dname = "dn"+str(noise_level)
    #
    print("###")
    print("dataset: ",dname)
    print("###")
    #
    data_genrator = SyntheticDataGeneratorBioinf(k, G, dict_entity_size, dict_mat_dtype,\
                                                list_mod_mat_id, sparsity_level, noise_level)
    data_dict = data_genrator.get_data_dict()
    #
    dict_name_dataset[dname] = data_dict

dict_name_dataset.keys()

#

fname = "./ncmf_sim_data/dict_name_dataset.pkl"
pkl.dump(dict_name_dataset,open(fname,"wb"))

# compressed_pickle(fname, dict_name_dataset)
