==> NCMF/ncmf_sim_study/5 - data_prep_cmf - 4 - final.ipynb <==
import pickle as pkl
import os
#
import pandas as pd
import numpy as np
import random
rng = 3
random.seed(rng)

fname = "./ncmf_sim_data/dict_name_dataset.pkl"

base_dname = "./ncmf_sim_data/"
base_dname_out = "./ncmf_sim_data/cmf/"

data_dict_all = pkl.load(open(fname,"rb"))

sample_id = 1
#
G = {"0":["0","1"],"1":["0","2"],"2":["0","3"],"3":["4","1"],"4":["5","2"]}
dict_entity_size = {'0':200, '1':500, '2':300, '3':400, '4':700, '5':600}
dict_mat_dtype = {'0':"real", '1':"real", '2':"binary", '3':"real", '4':"binary"}
target_matrix_index = [0,2]
#
data_dict_all_cmf_pp = {}

for dataset_name in data_dict_all.keys():    
    cur_dataset_dict_cmf_pp = {}
    data_folder = base_dname + dataset_name + "/"
    out_folder = base_dname_out + dataset_name + "/"
    if not os.path.exists(out_folder):
        os.makedirs(out_folder)
    print("data_folder: ",data_folder)
    print("out_folder: ",out_folder)
    print("###")
    #
    # Output files
    sampled_node_file = data_folder + f'sampled{sample_id}_node.dat'
    sampled_link_file = data_folder + f'sampled{sample_id}_link.dat'
    if target_matrix_index == -1:
        sampled_link_test_file = data_folder + f'sampled{sample_id}_link.dat.test'
    else:
        sampled_link_test_files = []
        for m in target_matrix_index:
            sampled_link_test_files.append(data_folder + f'sampled{sample_id}_link.dat.test.{m}')
    sampled_label_file = data_folder + f'sampled{sample_id}_label.dat'
    sampled_label_test_file = data_folder + f'sampled{sample_id}_label.dat.test'
    sampled_meta_file = data_folder + f'sampled{sample_id}_meta.dat'
    sampled_info_file = data_folder + f'sampled{sample_id}_info.dat'
    #
    sampled_dict_id_idx_file = data_folder + f"sampled{sample_id}_dict_id_idx.pkl" 
    #
    if not os.path.exists(data_folder):
        assert False
    #
    if not (os.path.isfile(sampled_node_file)\
           or os.path.isfile(sampled_link_file)\
           or os.path.isfile(sampled_label_file)\
           or os.path.isfile(sampled_label_test_file)\
           or os.path.isfile(sampled_meta_file)\
           or os.path.isfile(sampled_info_file)\
           or os.path.isfile(sampled_dict_id_idx_file)):
        assert False    
    #
    if target_matrix_index == -1:
        if not (os.path.isfile(sampled_link_test_file)):
            assert False
    else:
        for cur_file in sampled_link_test_files:
            if not (os.path.isfile(cur_file)):
                assert False
    #
    for cur_mat_idx in G.keys():
        cur_row_eid = G[cur_mat_idx][0]
        cur_col_eid = G[cur_mat_idx][1]
        cur_dataset_dict_cmf_pp[cur_mat_idx] = np.zeros((dict_entity_size[cur_row_eid],\
                                                         dict_entity_size[cur_col_eid]))
        print("cur_mat_idx: ",cur_mat_idx,\
              "cur_dataset_dict_cmf_pp[cur_mat_idx].shape: ",cur_dataset_dict_cmf_pp[cur_mat_idx].shape)
    print("#")
    #
    dict_id_idx_file = pkl.load(open(sampled_dict_id_idx_file,"rb"))
    #
    #populate the matrices
    df_link_train = pd.read_csv(sampled_link_file,sep="\t",header=None)
    df_link_train.columns = ["left", "right", "link_type", "value"]  
    #
    list_all_link_types_from_link_file = list(df_link_train["link_type"].unique())
    list_all_link_types_from_G = [int(x) for x in list(G.keys())]
    assert set(list_all_link_types_from_link_file) == set(list_all_link_types_from_G)
    #
    for cur_mat_id in list_all_link_types_from_link_file:
        df_link_train_subset = df_link_train[df_link_train["link_type"].isin([cur_mat_id])]
        cur_mat = cur_dataset_dict_cmf_pp[str(cur_mat_id)]
        print("cur_mat: ",cur_mat_id,", bef#: ",np.sum(cur_mat))
        val_leq_zero_count = 0
        for idx,row in df_link_train_subset.iterrows():
            left_id = int(row["left"])
            left_idx = dict_id_idx_file[left_id]
            right_id = int(row["right"])
            right_idx = dict_id_idx_file[right_id]
            value = row["value"]
            #
            cur_mat[left_idx,right_idx] = value
            #
            if not value > 0:
                val_leq_zero_count+=1
        print("cur_mat: ",cur_mat_id,", aft#: ",np.sum(cur_mat>0))        
        assert np.sum(cur_mat > 0) + val_leq_zero_count == df_link_train_subset.shape[0]
        print("#")
        #
        cur_fname = out_folder+str(cur_mat_id)+".csv"
        pd.DataFrame(cur_mat).to_csv(cur_fname,header=None,index=False)
        print("persisting: ",cur_fname)
        print("#")
    #
    data_dict_all_cmf_pp[dataset_name] = cur_dataset_dict_cmf_pp

data_dict_all_cmf_pp.keys()

data_dict_all_cmf_pp["dt1"].keys()

list_all_link_types_from_link_file
