==> NCMF/ncmf_sim_study/7 -  gcmf_sim_results_gen - try2 - final.ipynb <==
import numpy as np
np.random.seed(0)
import pandas as pd
import os
from datetime import datetime
import time
from sklearn.preprocessing import MaxAbsScaler, StandardScaler, maxabs_scale
from math import pi
from numpy import sin, cos, linspace
import random
import seaborn as sns
import pandas as pd
import matplotlib.pyplot as plt
import pickle as pkl

from sklearn.metrics import roc_auc_score, mean_squared_error

sample_no = 1
data_dir = "./ncmf_sim_data"
#list_dataset_names = ["dt1"]
#list_dataset_names = ["ds1","ds2","ds3"]
list_dataset_names = ['dt1', 'ds1', 'ds2', 'ds3', 'dn1', 'dn2', 'dn3']

matrix_types = {
    "real": ["X0","X1","X3"],
    "binary": ["X2","X4"]
}

list_test_mat_ids = ["X0","X2"]
list_test_mat_idx = [0,2]
#
# list_test_mat_ids = ["X0"]
# list_test_mat_idx = [0]

def get_perf_metrics(sample_no,\
                        data_dir,\
                        list_dataset_names,\
                        matrix_types,\
                        list_test_mat_ids,\
                        list_test_mat_idx
                        ):
    dict_dname_perf = {}
    out_dir_base = data_dir + "/gcmf/out/"
    #
    for dataset_name in list_dataset_names:
        print("dataset_name: ",dataset_name)
        dict_dname_perf[dataset_name] = {}
        #
        out_dir_recons = out_dir_base + dataset_name + "/"
        data_dir_gt = data_dir + "/" + dataset_name + "/"
        #
        sampled_dict_id_idx_file = data_dir_gt + f"sampled{sample_no}_dict_id_idx.pkl"
        dict_id_idx = pkl.load(open(sampled_dict_id_idx_file,"rb"))
        #
        for i in np.arange(len(list_test_mat_idx)):       
            test_mat_idx = list_test_mat_idx[i]
            test_mat_id = list_test_mat_ids[i]
            #
            fname_pred = out_dir_recons + str(int(test_mat_idx)+1) + "_pred.csv"
            X_pred = pd.read_csv(fname_pred,header=None).to_numpy()
            #
            fname_list_gt = data_dir_gt + "sampled1_link.dat.test."+str(test_mat_idx)
            df_list_gt = pd.read_csv(fname_list_gt,sep="\t",header=None)
            df_list_gt.columns = ["i_idx","j_idx","value"]
            #
            list_gt_vals = []
            list_pred_vals = []
            for idx,row in df_list_gt.iterrows():
                i = int(dict_id_idx[row["i_idx"]])
                j = int(dict_id_idx[row["j_idx"]])
                #
                cur_val_gt = row["value"]
                list_gt_vals.append(cur_val_gt)
                #
                cur_val_pred = X_pred[i,j]
                list_pred_vals.append(cur_val_pred)
            #
            print("mat_idx:",test_mat_idx,", list_gt_vals - min: ",np.min(list_gt_vals),", max: ",np.max(list_gt_vals))
            print("mat_idx:",test_mat_idx,", list_pred_vals - min: ",np.min(list_pred_vals),", max: ",np.max(list_pred_vals))
            #
            if test_mat_id in matrix_types["real"]:
                eval_mat_dtype = "real"
            elif test_mat_id in matrix_types["binary"]:
                eval_mat_dtype = "binary"
            else:
                assert False    
            #
            if eval_mat_dtype == "real":
                cur_perf = mean_squared_error(list_gt_vals,list_pred_vals)
                cur_perf_metric = "mse"
            elif eval_mat_dtype == "binary":
                cur_perf = roc_auc_score(list_gt_vals,list_pred_vals)
                cur_perf_metric = "auc"
            else:
                assert False
            #
            dict_dname_perf[dataset_name][test_mat_idx] = {
                "perf":cur_perf,
                "perf_metric":cur_perf_metric
            }
        print("###")
    #
    pkl.dump(dict_dname_perf,open(out_dir_base+"dict_perf_gcmf.pkl","wb"))
    #
    return dict_dname_perf

dict_perf = get_perf_metrics(sample_no,\
                        data_dir,\
                        list_dataset_names,\
                        matrix_types,\
                        list_test_mat_ids,\
                        list_test_mat_idx)

pd.DataFrame(dict_perf)

dict_perf
